refactor(foundation): move model parsing to separate file
